shared/pyproject.toml
[project]
name = "rag-shared"
version = "0.1.0"
dependencies = [
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp-proto-grpc",
    "opentelemetry-instrumentation-fastapi",
]

[tool.setuptools.packages.find]
where = ["."]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
